refactor: split types into multiple files
