+++ packages/types/integrations/messages.ts
@@ -1,0 +1,50 @@
+import { z } from 'zod';
+
+export const messageGetSchema = z.object({
+  messageId: z.string().uuid(),
+});
+export type messageGetType = z.infer<typeof messageGetSchema>;
+export type messageGetResponseType = {
+  body: string;
+  author: {
+    displayName: string | null;
+  } | null;
+  channelId: string;
+  externalMessageId: string | null;
+  threadId: string | null;
+} | null;
+
+export const messageFindSchema = z.object({
+  channelId: z.string().uuid(),
+  externalMessageId: z.string().min(1).optional(),
+  threadId: z.string().uuid().optional(),
+  where: z
+    .object({
+      sort: z.enum(['sentAt']).optional(),
+      order: z.enum(['asc', 'desc']).optional(),
+      externalMessageId: z.boolean().optional(),
+    })
+    .optional(),
+});
+export type messageFindType = z.infer<typeof messageFindSchema>;
+export type messageFindResponseType = {
+  externalMessageId: string | null;
+  threadId: string | null;
+  id: string;
+} | null;
+
+export const messagePostSchema = z.object({
+  accountId: z.string().uuid(),
+  channelId: z.string().uuid(),
+  threadId: z.string().uuid(),
+  authorId: z.string().uuid(),
+  externalMessageId: z.string().min(1),
+  body: z.string().min(1),
+});
+export type messagePostType = z.infer<typeof messagePostSchema>;
+
+export const messagePutSchema = z.object({
+  messageId: z.string().uuid(),
+  externalMessageId: z.string().min(1).optional(),
+});
+export type messagePutType = z.infer<typeof messagePutSchema>;
