+++ packages/types/integrations/users.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+export const userPostSchema = z.object({
+  externalUserId: z.string().min(1),
+  accountsId: z.string().uuid(),
+  displayName: z.string().min(1),
+  profileImageUrl: z.string().min(1).optional(),
+});
+export type userPostType = z.infer<typeof userPostSchema>;
