+++ packages/types/package.json
@@ -5,11 +5,12 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "dev": "tsc --watch",
-    "build": "tsc"
+    "dev": "tsup --watch",
+    "build": "tsup"
   },
   "devDependencies": {
-    "typescript": "^4.9.3"
+    "typescript": "^4.9.3",
+    "tsup": "^6.5.0"
   },
   "dependencies": {
     "zod": "^3.20.2"
