+++ packages/types/tsup.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+  entry: ['index.ts'],
+  splitting: false,
+  clean: true,
+  dts: true,
+});
